+++ tracker/test/support/html-fixtures.ts
@@ -1,0 +1,6 @@
+/**
+ * This is a stub for custom form onsubmit handlers Plausible users may have on their websites.
+ * Overriding onsubmit with a custom handler is common practice in web development for a variety of reasons (mostly UX),
+ * so this stub can be used in tests to simulate such behavior.
+ */
+export const customSubmitHandlerStub = "event.preventDefault(); console.log('Submitted')"